+++ record_platform_interface/lib/src/types/types.dart
@@ -6,5 +6,6 @@ export 'input_device.dart';
 export 'input_device_type.dart';
 export 'ios_audio_session.dart';
 export 'ios_record_config.dart';
+export 'pcm_format.dart';
 export 'record_config.dart';
 export 'record_state.dart';